adding to the employeeStock (dictionary) in EmployeeSystem class.

+++ Degree.cs
@@ -7,9 +7,9 @@ namespace MamaLis
     class Degree
     {
         private Dictionary<string, bool> _degreesDictionary = new Dictionary<string, bool>();
-        private float _extraRisk;
+        private double _extraRisk;
 
-        public Degree(bool ifExpetrt, bool ifMakeDecisions, float extrarisk)
+        public Degree(bool ifExpetrt, bool ifMakeDecisions, double extrarisk)
         {
             this._degreesDictionary.Add("Expert", ifExpetrt);
             this._degreesDictionary.Add("Make Decisions", ifMakeDecisions);
